+++ tinper-bee/bee-radio/CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="2.0.5"></a>
+## [2.0.5](https://github.com/tinper-bee/bee-radio/compare/v2.0.4...v2.0.5) (2019-04-04)
+
+
+
 <a name="2.0.4"></a>
 ## [2.0.4](https://github.com/tinper-bee/bee-radio/compare/v2.0.3...v2.0.4) (2019-04-01)